+++ CHANGELOG.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+Release 1.1.0
+-------------
+
+- Support for new policy scheme
+- Support for namspace policies
+- Add brew install command for MacOSX
+- Add pagination for dataset export command
+
 Release 1.0.0
 -------------